style: calendar dark mode friendly (#406)
* style: 🎨 shadcn primary to pink

* chore: 🔧 rm duplicate figtree import
