+++ src/components/creation/calendar/calendar-body-day-cell.tsx
@@ -1,3 +1,4 @@
+import { Typography } from "@mui/material";
 import { cn } from "@/lib/utils";
 import type { ZotDate } from "@/lib/zotdate";
 
@@ -14,21 +15,23 @@ export function CalendarBodyDayCell({
 }: CalendarBodyDayCellProps) {
 	const isSelected = calendarDay.isSelected;
 
+	const dayClassName = cn(
+		"relative aspect-square h-10 w-10 flex-center rounded-xl text-2xl",
+		isSelected && "bg-primary text-gray-50",
+		isHighlighted && "bg-slate-base text-gray-dark",
+	);
+
 	return (
-		<p
-			className={cn(
-				"relative aspect-square h-6 w-6 flex-center rounded-full font-medium text-base text-gray-dark text-sm md:h-8 md:w-8 md:rounded-xl md:text-base",
-				isSelected && "bg-primary text-gray-50",
-				isHighlighted && "bg-slate-base text-gray-dark",
-				!isCurrentMonth &&
-					cn(
-						"text-gray-base",
-						isHighlighted && "bg-opacity-30",
-						isSelected && "bg-opacity-50 text-gray-100",
-					),
+		<div>
+			{isCurrentMonth ? (
+				<Typography color="textPrimary" className={dayClassName} variant="h6">
+					{calendarDay.getDay()}
+				</Typography>
+			) : (
+				<Typography color="textDisabled" className={dayClassName} variant="h6">
+					{calendarDay.getDay()}
+				</Typography>
 			)}
-		>
-			{calendarDay.getDay()}
-		</p>
+		</div>
 	);
 }
