+++ src/fonts.ts
@@ -1,0 +1,8 @@
+import { Figtree } from "next/font/google";
+
+export const figtree = Figtree({
+	subsets: ["latin"],
+	variable: "--font-figtree",
+	display: "swap",
+	weight: ["400", "500", "600", "700"],
+});
